Megurin UI Added, Animation Snap Code Added
Shifted Megurins animations to UI and setup character on Marketing Screen

As for the issues with Lady Parhelion's height, continued with plan of adding masks for all the buttons. Created a script with late update to force animations back down into acceptable limits down on the buttons. Hardcoded a few exceptions for Lady Parhelion's animations that can be expanded